requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio-mqtt==0.12.1
attrs==22.1.0
certifi==2024.6.2
charset-normalizer==2.1.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
paho-mqtt==1.6.1
requests==2.28.1
six==1.16.0
urllib3==1.26.19
websocket-server==0.6.4
websockets==12.0
yarl==1.9.4